# api/docs/swagger.yaml
definitions:
...
  payment.CreatePaymentReq:
    properties:
      amount:
        description: Decimal as a string to avoid precision loss
        type: string
      status:
        description: '''pending'', ''completed'', etc.'
        type: string
    type: object
  payment.CreatePaymentRes:
    properties:
      id:
        description: Payment ID
        type: string
    type: object
  payment.DeletePaymentRes:
    properties:
      success:
        type: boolean
    type: object
  payment.GetAllPaymentRes:
    properties:
      payments:
        items:
          $ref: '#/definitions/payment.Payment'
        type: array
    type: object
  payment.GetPaymentRes:
    properties:
      payment:
        $ref: '#/definitions/payment.Payment'
    type: object
  payment.Payment:
    properties:
      amount:
        description: Decimal as a string
        type: string
      id:
        type: string
      status:
        description: '''pending'', ''completed'', etc.'
        type: string
      transaction_date:
        type: string
    type: object
  report.CreateReportRequest:
    properties:
      description:
        type: string
      email:
        type: string
      problem:
        type: string
      property_id:
        type: string
      user_id:
        type: string
    type: object
  report.CreateReportResponse:
    properties:
      report:
        $ref: '#/definitions/report.Report'
    type: object
  report.GetReportResponse:
    properties:
      report:
        $ref: '#/definitions/report.Report'
    type: object
  report.Report:
    properties:
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      problem:
        type: string
      property_id:
        type: string
      user_id:
        type: string
    type: object
  report.Void:
    type: object
  request.CreateRequestRequest:
    properties:
      email:
        items:
          type: string
        type: array
      phone_number:
        type: string
      roommate_count:
        type: integer
      user_id:
        type: string
    type: object
  request.CreateRequestResponse:
    properties:
      request:
        $ref: '#/definitions/request.Request'
    type: object
  request.GetRequestResponse:
    properties:
      request:
        $ref: '#/definitions/request.Request'
    type: object
  request.Request:
    properties:
      email:
        description: email array
        items:
          type: string
        type: array
      id:
        type: string
      phone_number:
        type: string
      roommate_count:
        type: integer
      user_id:
        type: string
    type: object
  request.Void:
    type: object
  reviews.CreateReviewReq:
    properties:
      comment:
        type: string
      property_id:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
  reviews.CreateReviewRes:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      property_id:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  reviews.DeleteReviewRes:
    properties:
      message:
        type: string
    type: object
  reviews.GetAllReviewRes:
    properties:
      review:
        items:
          $ref: '#/definitions/reviews.Review'
        type: array
    type: object
  reviews.GetByIdReviewRes:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      property_id:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  reviews.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      property_id:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for Api-gateway Service
  title: Api Gateway
  version: "1.0"
paths:
...
  /payment/createpayment:
    post:
      consumes:
      - application/json
      description: Create a new payment with specified details
      parameters:
      - description: Create Payment Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payment.CreatePaymentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.CreatePaymentRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new payment
      tags:
      - payments
  /payment/deletepayment/{payment_id}:
    delete:
      consumes:
      - application/json
      description: Delete a payment by payment ID
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.DeletePaymentRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete payment
      tags:
      - payments
  /payment/getallpayment/{limit}/{page}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all payments with pagination
      parameters:
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.GetAllPaymentRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all payments
      tags:
      - payments
  /payment/getbyidpayment/{payment_id}:
    get:
      consumes:
      - application/json
      description: Retrieve payment details by payment ID
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.GetPaymentRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get payment by ID
      tags:
      - payments
...
  /report/create:
    post:
      consumes:
      - application/json
      description: Add a new user report
      parameters:
      - description: Create Report Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/report.CreateReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.CreateReportResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new report
      tags:
      - report
  /report/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.Void'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a report
      tags:
      - report
  /report/get/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.GetReportResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a report by ID
      tags:
      - report
  /request/create:
    post:
      consumes:
      - application/json
      description: Add a new user request
      parameters:
      - description: Create Request Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.CreateRequestResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new request
      tags:
      - request
  /request/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a request by its ID
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.Void'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a request
      tags:
      - request
  /request/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific request by its ID
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.GetRequestResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a request by ID
      tags:
      - request
  /review/create:
    post:
      consumes:
      - application/json
      description: Create a new review with specified details
      parameters:
      - description: Create Review Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/reviews.CreateReviewReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reviews.CreateReviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new review
      tags:
      - reviews
  /review/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reviews.DeleteReviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a review
      tags:
      - reviews
  /review/getallreview:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all reviews
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reviews.GetAllReviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all reviews
      tags:
      - reviews
  /review/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: Get the details of a specific review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reviews.GetByIdReviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a review by ID
      tags:
      - reviews
  /review/update/{id}:
    put:
      consumes:
      - application/json
      description: Update the details of a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Review Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/reviews.CreateReviewReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reviews.CreateReviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update an existing review
      tags:
      - reviews
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"